backend/requirements.txt
alembic==1.15.1
anthropic==0.49.0
bcrypt==4.3.0
click==8.1.8
cryptography==44.0.2
distro==1.9.0
fastapi==0.115.11
jiter==0.9.0
jsonpointer==2.0
langchain-anthropic==0.3.10
langchain-openai==0.3.9
langchain-tavily==0.1.5
langgraph==0.3.18
langgraph-checkpoint==2.0.21
langgraph-prebuilt==0.1.4
langgraph-sdk==0.1.58
Mako==1.3.9
MarkupSafe==3.0.2
msgpack==1.1.0
mypy==1.15.0
mypy-extensions==1.0.0
openai==1.68.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
regex==2024.11.6
starlette==0.46.1
tavily-python==0.5.1
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
uvicorn==0.34.0
